QR code generator with example and steps
QR codes are used to encode and decode the data into a machine-readable form. The use of camera phones to read two-dimensional barcodes for various purposes is currently a popular topic in both types of research and practical applications. I used this to make a QR  code for my GitHub profile.